requirements.txt
streamlit
sentence-transformers
umap-learn
matplotlib
scikit-learn
numpy